refactor: extract a protocol-agnostic graphics layer from kitty
Move the terminal-graphics abstraction out of the kitty module so a second
protocol can plug in: a GraphicsRenderer trait, a GraphicsManager (generalized
from KittyManager) owning image-id tracking and the erase lifecycle, and shared
helpers (tmux/escape sequences, image data, fit math) under a new graphics
module. KittyImageRenderer now implements GraphicsRenderer. No behavior change.
